выполнение задания 5.12. Модуляция
