+++ index.html
@@ -251,7 +251,9 @@
       <button type="button" class="error__button">Попробовать снова</button>
     </div>
   </template>
-  <script src="js/main.js"></script>
+  <script src="js/main.js" type="module"></script>
+  <script src="js/data.js" type="module"></script>
+  <script src="js/setup.js" type="module"></script>
 </body>
 
 </html>
