+++ js/data.js
@@ -1,0 +1,39 @@
+const TYPES = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
+const TIME = ['12:00', '13:00', '14:00'];
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+const PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+const DESCRIPTION = ['Описание помещения',
+  'Великолепная квартира-студия',
+  'Гнездышко для молодоженов',
+  'Квартира для студентов',
+  'Для многодетной семьи',
+  'Для одиночек с животными',
+];
+
+const MAX_ADS = 10;
+const MIN_PRICE = 1;
+const MAX_PRICE = 30000;
+const MIN_ROOMS = 1;
+const MAX_ROOMS = 5;
+const MIN_GUESTS = 1;
+const MAX_GUESTS = 10;
+const MIN_LAT = 35.65000;
+const MAX_LAT = 35.70000;
+const MIN_LNG = 139.70000;
+const MAX_LNG = 139.80000;
+const COORD_PRECISION = 5;
+const PARAMS = [MAX_ADS, MIN_PRICE, MAX_PRICE, MIN_ROOMS, MAX_ROOMS, MIN_GUESTS, MAX_GUESTS, MIN_LAT, MAX_LAT, MIN_LNG, MAX_LNG, COORD_PRECISION,
+];
+
+export { TYPES, TIME, FEATURES, PHOTOS, DESCRIPTION, PARAMS };
